add support for paraformer / zipformer2_ctc models
and add a bunch of pretrained models to the list, including nvidia nemo
conformer hybrid transducer/ctc

unfortunately, only the transducer variants are supported as of now in
the C api of sherpa-onnx in streaming mode. sherpa-onnx does have the
logic there to run them but the config parameters aren't passed through
yet. i may get around to implementing that in upstream at some point.

specifically, the model arch we'd need is nemo_ctc in online mode.

with that said, transducers are better anyway so it's NBD :]

closes #1